Reorganize scripts + a few cosmetic changes
Introduce bat/sh wrapper scripts for easier usage.
Split make.py into Build.py and Run.py.
Use CamelCase for Python class names.
Update .gitignore.
