+++ cmake/build_configuration.cmake
@@ -6,5 +6,5 @@ string(TOUPPER ${CMAKE_BUILD_TYPE} upper_build_type)
 set(cflags_var "CMAKE_C_FLAGS_${upper_build_type}")
 set(cxxflags_var "CMAKE_CXX_FLAGS_${upper_build_type}")
 
-message("C++ compiler flags: ${${cxxflags_var}}")
-message("C compiler flags: ${${cflags_var}}")
+message(STATUS "C++ compiler flags: ${${cxxflags_var}}")
+message(STATUS "C compiler flags: ${${cflags_var}}")
