+++ run.bat
@@ -1,0 +1,1 @@
+python3 .\scripts\Run.py %*
